Spec text update on homepage
The original rfc7033 spec, evidently, contained an error ("for example,
the that a") which made me squint at this paragraph a bit. Looking at
the official spec on the IETF website, it seems like there were a couple
updates to the quote that currently appears on webfinger.net, this patch
updates the quote from the spec to its most recent version (even though
I think changing "include" to "includes" below is not grammatically
correct).